# Release v0.5.8b 2026-08-09
## New Features

1. `Info memory` command implmented
2. Deep test on globs
3. glob perf tests verifying 8x improvement
4. RESP3 support
5. add `CONFIG GET`
6. add 'CLIENT LIST'
7. new lzr_log api (initialazer logging)
8. new sharding config for later
9. new docs website

## Fixes

1. Array serialization inconsistencies
2. Out of bounds issue in glob code
3. key length error for globs
4. correct numeric behavior
5. fix some resp2 hello support
6. RESP3 support
7. fix array interface
8. fix some config conversion issues
9. add diagnostics for perftest
10. fix asynch calls probably
11. refactor keys api
12. fix potential bugs with query_key
13. fix more hidden state potential bugs
14.sharding refactor
15. ci git issue
16. refactor rest of API's
17. update TODO
18. update .gitignore
19. improve ordered sharding (claude)
20. fix older gcc build issue
21. fix todo 38. (Many OrderedSet interface issues)

+++ src/keys_api.cpp
@@ -360,7 +360,10 @@ static int BarchModifyDouble(caller& call,const arg_t& argv, double by) {
 )
 int INCR(caller& call, const arg_t& argv) {
     ++statistics::incr_ops;
-    return BarchModifyInteger(call, argv, 1);
+    // (int64_t) matters: a bare 1 deduces IntT as int, so INCR ran on 32 bits and
+    // refused any value above INT32_MAX - redis's INCR is 64 bit signed. INCRBY was
+    // always right because it parses its argument into a long long. See DONE 39
+    return BarchModifyInteger(call, argv, (int64_t) 1);
 }
 int cmd_INCR(ValkeyModuleCtx *ctx, ValkeyModuleString **argv, int argc) {
     vk_caller call;
@@ -494,7 +497,7 @@ int cmd_UINCRBY(ValkeyModuleCtx *ctx, ValkeyModuleString **argv, int argc) {
 
 int DECR(caller& call, const arg_t& argv) {
     ++statistics::decr_ops;
-    return BarchModifyInteger(call, argv, -1);
+    return BarchModifyInteger(call, argv, (int64_t) -1);
 }
 
 int cmd_DECR(ValkeyModuleCtx *ctx, ValkeyModuleString **argv, int argc) {
